+++ src/cmd/commands.cpp
@@ -167,9 +167,18 @@ void Commands::evaluate(const std::string& path) {
   Program program = SequenceProgram::getProgramAndSeqId(path).first;
   Evaluator evaluator(settings, EVAL_ALL, false);
   Sequence seq;
-  evaluator.eval(program, seq);
+  std::string error;
+  try {
+    evaluator.eval(program, seq);
+  } catch (const std::exception& e) {
+    error = e.what();
+  }
   if (!settings.print_as_b_file) {
     std::cout << seq << std::endl;
+  }
+  if (!error.empty()) {
+    std::cout << "Error: " << error << std::endl;
+    exit(1);
   }
 }